Implementing MultipartReplyDesc handler (not tested)
It will be necessary to change controller design to prevent memory leak. Switches rack will be implemented only in the application side because if application try to access a switch that controller already removed from de container, it will break controller. Application will hold a pointer to the connection and will check it before to do anything. When controller advise about a switch down event, application drops the switch from de rack.