Supporting more than one switches. Locks and unlocks properly used.
Switches objects are stacked in a virtual rack; untill now it was not
concurrency safe causing program abort when the second switch connects.
Fluid base provides lock and unlock methods to ensure safety resource
access; these methods are in used every time the stack is handled.

It was tested and it is working properly.

Next steps are provide some missing logs and to spawn application
methods called by controller in threads.