Learning the source (Building and working properly).
Controller is checking every packet-in; it extracts information about
the source device, such as MAC, IP and the port device is connected to
the switch. If device is known (it means device information is stored)
the entry is updated in switch's arp table; if not a new entry is added.

It was testes only with IPv4.